+++ ckanext/dcat/profiles/euro_dcat_ap_2.py
@@ -1,4 +1,5 @@
 import json
+from decimal import Decimal, DecimalException
 
 from rdflib import URIRef, BNode, Literal
 from ckanext.dcat.utils import resource_uri
@@ -11,6 +12,7 @@ from .base import (
     DCATAP,
     DCT,
     XSD,
+    SCHEMA,
 )
 
 from .euro_dcat_ap import EuropeanDCATAPProfile
@@ -31,9 +33,13 @@ class EuropeanDCATAP2Profile(EuropeanDCATAPProfile):
         # call super method
         super(EuropeanDCATAP2Profile, self).parse_dataset(dataset_dict, dataset_ref)
 
+        # Standard values
+        value = self._object_value(dataset_ref, DCAT.temporalResolution)
+        if value:
+            dataset_dict["extras"].append({"key": "temporal_resolution", "value": value})
+
         # Lists
         for key, predicate in (
-            ("temporal_resolution", DCAT.temporalResolution),
             ("is_referenced_by", DCT.isReferencedBy),
             ("applicable_legislation", DCATAP.applicableLegislation),
             ("hvd_category", DCATAP.hvdCategory),
@@ -54,14 +60,20 @@ class EuropeanDCATAP2Profile(EuropeanDCATAPProfile):
             self._add_spatial_to_dict(dataset_dict, key, spatial)
 
         # Spatial resolution in meters
-        spatial_resolution_in_meters = self._object_value_float_list(
+        spatial_resolution = self._object_value_float_list(
             dataset_ref, DCAT.spatialResolutionInMeters
         )
-        if spatial_resolution_in_meters:
+        if spatial_resolution:
+            # For some reason we incorrectly allowed lists in this property at some point
+            # keep support for it but default to single value
+            value = (
+                spatial_resolution[0] if len(spatial_resolution) == 1
+                else json.dumps(spatial_resolution)
+            )
             dataset_dict["extras"].append(
                 {
                     "key": "spatial_resolution_in_meters",
-                    "value": json.dumps(spatial_resolution_in_meters),
+                    "value": value,
                 }
             )
 
@@ -147,15 +159,17 @@ class EuropeanDCATAP2Profile(EuropeanDCATAPProfile):
             dataset_dict, dataset_ref
         )
 
+        # Standard values
+        self._add_triple_from_dict(
+            dataset_dict,
+            dataset_ref,
+            DCAT.temporalResolution,
+            "temporal_resolution",
+            _datatype=XSD.duration,
+        )
+
         # Lists
         for key, predicate, fallbacks, type, datatype in (
-            (
-                "temporal_resolution",
-                DCAT.temporalResolution,
-                None,
-                Literal,
-                XSD.duration,
-            ),
             ("is_referenced_by", DCT.isReferencedBy, None, URIRefOrLiteral, None),
             (
                 "applicable_legislation",
@@ -178,6 +192,14 @@ class EuropeanDCATAP2Profile(EuropeanDCATAPProfile):
             )
 
         # Temporal
+
+        # The profile for DCAT-AP 1 stored triples using schema:startDate,
+        # remove them to avoid duplication
+        for temporal in self.g.objects(dataset_ref, DCT.temporal):
+            if SCHEMA.startDate in [t for t in self.g.predicates(temporal, None)]:
+                self.g.remove((temporal, None, None))
+                self.g.remove((dataset_ref, DCT.temporal, temporal))
+
         start = self._get_dataset_value(dataset_dict, "temporal_start")
         end = self._get_dataset_value(dataset_dict, "temporal_end")
         if start or end:
@@ -216,10 +238,10 @@ class EuropeanDCATAP2Profile(EuropeanDCATAPProfile):
                         (
                             dataset_ref,
                             DCAT.spatialResolutionInMeters,
-                            Literal(float(value), datatype=XSD.decimal),
+                            Literal(Decimal(value), datatype=XSD.decimal),
                         )
                     )
-                except (ValueError, TypeError):
+                except (ValueError, TypeError, DecimalException):
                     self.g.add(
                         (dataset_ref, DCAT.spatialResolutionInMeters, Literal(value))
                     )
@@ -278,7 +300,7 @@ class EuropeanDCATAP2Profile(EuropeanDCATAPProfile):
                     ("license", DCT.license, None, URIRefOrLiteral),
                     ("access_rights", DCT.accessRights, None, URIRefOrLiteral),
                     ("title", DCT.title, None, Literal),
-                    ("endpoint_description", DCAT.endpointDescription, None, Literal),
+                    ("endpoint_description", DCAT.endpointDescription, None, URIRefOrLiteral),
                     ("description", DCT.description, None, Literal),
                 ]