+++ ckanext/dcat/tests/test_scheming_support.py
@@ -1,5 +1,6 @@
 from unittest import mock
 import json
+from decimal import Decimal
 
 import pytest
 from rdflib.namespace import RDF
@@ -19,7 +20,6 @@ from ckanext.dcat.profiles import (
     XSD,
     VCARD,
     FOAF,
-    SCHEMA,
     SKOS,
     LOCN,
     GSP,
@@ -101,7 +101,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
                 {"start": "1905-03-01", "end": "2013-01-05"},
                 {"start": "2024-04-10", "end": "2024-05-29"},
             ],
-            "temporal_resolution": ["PT15M", "P1D"],
+            "temporal_resolution": "PT15M",
             "spatial_coverage": [
                 {
                     "geom": {
@@ -133,7 +133,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
                     "centroid": {"type": "Point", "coordinates": [1.26639, 41.12386]},
                 }
             ],
-            "spatial_resolution_in_meters": [1.5, 2.0],
+            "spatial_resolution_in_meters": 1.5,
             "resources": [
                 {
                     "name": "Resource 1",
@@ -195,6 +195,13 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(g, dataset_ref, DCT.type, dataset["dcat_type"])
         assert self._triple(g, dataset_ref, ADMS.versionNotes, dataset["version_notes"])
         assert self._triple(g, dataset_ref, DCT.accessRights, dataset["access_rights"])
+        assert self._triple(
+            g,
+            dataset_ref,
+            DCAT.spatialResolutionInMeters,
+            dataset["spatial_resolution_in_meters"],
+            data_type=XSD.decimal,
+        )
 
         # Dates
         assert self._triple(
@@ -210,6 +217,13 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
             DCT.modified,
             dataset["modified"],
             data_type=XSD.date,
+        )
+        assert self._triple(
+            g,
+            dataset_ref,
+            DCAT.temporalResolution,
+            dataset["temporal_resolution"],
+            data_type=XSD.duration,
         )
 
         # List fields
@@ -232,23 +246,12 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
             == dataset["documentation"]
         )
         assert (
-            self._triples_list_values(g, dataset_ref, DCAT.temporalResolution)
-            == dataset["temporal_resolution"]
-        )
-        assert (
             self._triples_list_values(g, dataset_ref, DCT.isReferencedBy)
             == dataset["is_referenced_by"]
         )
         assert (
             self._triples_list_values(g, dataset_ref, DCATAP.applicableLegislation)
             == dataset["applicable_legislation"]
-        )
-
-        assert (
-            self._triples_list_python_values(
-                g, dataset_ref, DCAT.spatialResolutionInMeters
-            )
-            == dataset["spatial_resolution_in_meters"]
         )
 
         # Repeating subfields
@@ -306,28 +309,28 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(
             g,
             temporal[0][2],
-            SCHEMA.startDate,
+            DCAT.startDate,
             dataset_dict["temporal_coverage"][0]["start"],
             data_type=XSD.date,
         )
         assert self._triple(
             g,
             temporal[0][2],
-            SCHEMA.endDate,
+            DCAT.endDate,
             dataset_dict["temporal_coverage"][0]["end"],
             data_type=XSD.date,
         )
         assert self._triple(
             g,
             temporal[1][2],
-            SCHEMA.startDate,
+            DCAT.startDate,
             dataset_dict["temporal_coverage"][1]["start"],
             data_type=XSD.date,
         )
         assert self._triple(
             g,
             temporal[1][2],
-            SCHEMA.endDate,
+            DCAT.endDate,
             dataset_dict["temporal_coverage"][1]["end"],
             data_type=XSD.date,
         )
@@ -340,15 +343,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
             g, spatial[0][2], SKOS.prefLabel, dataset["spatial_coverage"][0]["text"]
         )
 
-        assert len([t for t in g.triples((spatial[0][2], LOCN.geometry, None))]) == 2
-        # Geometry in GeoJSON
-        assert self._triple(
-            g,
-            spatial[0][2],
-            LOCN.geometry,
-            dataset["spatial_coverage"][0]["geom"],
-            GEOJSON_IMT,
-        )
+        assert len([t for t in g.triples((spatial[0][2], LOCN.geometry, None))]) == 1
         # Geometry in WKT
         wkt_geom = wkt.dumps(dataset["spatial_coverage"][0]["geom"], decimals=4)
         assert self._triple(g, spatial[0][2], LOCN.geometry, wkt_geom, GSP.wktLiteral)
@@ -404,7 +399,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         )
 
         assert self._triple(
-            g, distribution_ref, DCAT.byteSize, float(resource["size"]), XSD.decimal
+            g, distribution_ref, DCAT.byteSize, Decimal(resource["size"]), XSD.decimal
         )
         # Checksum
         checksum = self._triple(g, distribution_ref, SPDX.checksum, None)[2]
@@ -621,7 +616,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(
             g,
             temporal[0][2],
-            SCHEMA.endDate,
+            DCAT.endDate,
             dataset_dict["temporal_coverage"][0]["end"],
             data_type=XSD.date,
         )
@@ -634,7 +629,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(
             g,
             temporal[0][2],
-            SCHEMA.startDate,
+            DCAT.startDate,
             dataset_dict["temporal_coverage"][0]["start"],
             data_type=XSD.dateTime,
         )
@@ -646,7 +641,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(
             g,
             temporal[1][2],
-            SCHEMA.startDate,
+            DCAT.startDate,
             dataset_dict["temporal_coverage"][1]["start"],
             data_type=XSD.dateTime,
         )
@@ -659,7 +654,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(
             g,
             temporal[2][2],
-            SCHEMA.startDate,
+            DCAT.startDate,
             "2024-11-24T00:00:00",
             data_type=XSD.dateTime,
         )
@@ -670,7 +665,7 @@ class TestSchemingSerializeSupport(BaseSerializeTest):
         assert self._triple(
             g,
             temporal[2][2],
-            SCHEMA.endDate,
+            DCAT.endDate,
             "2012-06-12T00:00:00",
             data_type=XSD.dateTime,
         )
@@ -767,6 +762,8 @@ class TestSchemingParseSupport(BaseParseTest):
 
         assert dataset["issued"] == u"2012-05-10"
         assert dataset["modified"] == u"2012-05-10T21:04:00"
+        assert dataset["temporal_resolution"] == "PT15M"
+        assert dataset["spatial_resolution_in_meters"] == "1.5"
 
         # List fields
         assert sorted(dataset["conforms_to"]) == ["Standard 1", "Standard 2"]
@@ -784,14 +781,7 @@ class TestSchemingParseSupport(BaseParseTest):
             "http://dataset.info.org/doc1",
             "http://dataset.info.org/doc2",
         ]
-        assert sorted(dataset["temporal_resolution"]) == [
-            "P1D",
-            "PT15M",
-        ]
-        assert sorted(dataset["spatial_resolution_in_meters"]) == [
-            1.5,
-            2.0,
-        ]
+
         assert sorted(dataset["is_referenced_by"]) == [
             "https://doi.org/10.1038/sdata.2018.22",
             "test_isreferencedby",